+++ CHANGELOG.rst
@@ -20,6 +20,8 @@ Common
 UC
 --
 
+    * fix: Modbus/UDP (server) packet processing
+
     * data pullers
     * auto_off unit state processor can be cancelled / modified while running